2.10-dev
[ -- Updated -- ] 
* AccountsEntity.java double{ #rating } to double[]{ #rating } variable. 
* AccountsDA.java refelect changes of AccountsEntity.java double[]{ #rating } changes. 
* CommentsEntity.java double{ #rating } to int{ #rating }. 
* CommentsDA.java reflects changes of CommentsEntity.java int{ #rating } changes. 
* CommentsViewController.java double{ #rating } to int{ #rating } to reflect changes of CommentsEntity.java int{ #rating } changes. 
* EditProfileViewController.java fix issue when favSpotComboBox is not selected upon saving. 
* Re-biuld accounts.db
* Re-build comments.db
 
[ -- Removed -- ] 
- AccountsEntity.java removed int{ #noRate }, not needed for new rating formula. 

[ -- Notes -- ] 
* accounts.db have been re-build due to AccountsEntity.java rating format change. Please re-add / re-register required account(s). 
* Please check all your classes that you are working on, that AccountsEntity.java didn't heavily impact your work. If it does, please make changes accordingly. 
* AccountsEntity.java has double{ #rating } variable changed to double[]{ #rating } instead. And has a size check of a fixed length of 5. 
* AccountsEntity.java has int{ #noRating } variable has been removed, as it isn't need for the new rating formula method. 

{ - How Should / Does The New Rating Work? - } 
User's rating on other accounts are only allowed the option to give 1 / 2 / 3 / 4 / 5 stars, but stars can still be displayed in 1 / 1.5 / 2 / 2.5 / 3 / 3.5 / 4 / 4.5 / 5 stars with weighted average formula.
Rating formula method has been changed from average calculation to weighted average calculation. You can read more about weighted average for 5 Stars rating on:
https://stackoverflow.com/questions/10196579/algorithm-used-to-calculate-5-star-ratings

+++ src/dataAccess/CommentsDA.java
@@ -23,13 +23,10 @@ public class CommentsDA {
 
 		comments = db.getTreeMap("comments");
 
-		/*
-		 * List<CommentsEntity> ceList = new ArrayList<>(); ceList.add(new
-		 * CommentsEntity("test", "test", "test", 2.5)); ceList.add(new
-		 * CommentsEntity("test2", "test2", "test2", 3.5));
-		 *
-		 * comments.put("admin", ceList);
-		 */
+		// List<CommentsEntity> ceList = new ArrayList<>();
+		// ceList.add(new CommentsEntity("test", "test", "test", 1));
+		// ceList.add(new CommentsEntity("test2", "test2", "test2", 4));
+		// comments.put("admin", ceList);
 		db.commit();
 	}
 
@@ -52,7 +49,7 @@ public class CommentsDA {
 		return data;
 	}
 
-	public static void addComment(String adminNo, String comment, double rating) {
+	public static void addComment(String adminNo, String comment, int rating) {
 		String sessionAdminNo = AccountsDA.getAdminNo();
 		String sessionName = AccountsDA.getName();
 		List<CommentsEntity> commentsList;
